Refactor: Break out uglify-js and clean-css in separate namespaces
This change focuses the code in optimus.optizations.minify on the mechanics of
minifying assets without being specific about what tool does that job. The idea
is to make it easier to provide alternatives.
